docs: Rewrite the Chinese (Traditional) translation
The quality of the original translation was inadequate –
hard-to-understand sentences, misleading expressions,
and some weird typos (for example, 奈秒.)

This commit refactors the translation, aiming to
be able to read easily, understand quickly, and
less “misleading” section.

This commit also converts the terms of
this translation to what the Taiwanese prefers
to say with https://zhconvert.org.

Signed-off-by: pan93412 <pan93412@gmail.com>